+++ InfiniTAM/ITMLib/Engine/ITMVisualisationEngine.cpp
@@ -20,7 +20,7 @@ void IITMVisualisationEngine::DepthToUchar4(ITMUChar4Image *dst, ITMFloatImage *
 {
 	Vector4u *dest = dst->GetData(MEMORYDEVICE_CPU);
 	float *source = src->GetData(MEMORYDEVICE_CPU);
-	int dataSize = dst->dataSize;
+	int dataSize = static_cast<int>(dst->dataSize);
 
 	memset(dst->GetData(MEMORYDEVICE_CPU), 0, dataSize * 4);
 
@@ -60,7 +60,7 @@ void IITMVisualisationEngine::NormalToUchar4(ITMUChar4Image *dst, ITMFloat4Image
 {
 	Vector4u *dest = dst->GetData(MEMORYDEVICE_CPU);
 	Vector4f *source = src->GetData(MEMORYDEVICE_CPU);
-	int dataSize = dst->dataSize;
+	int dataSize = static_cast<int>(dst->dataSize);
 
 	memset(dst->GetData(MEMORYDEVICE_CPU), 0, dataSize * 4);
 	{
@@ -83,7 +83,7 @@ void IITMVisualisationEngine::WeightToUchar4(ITMUChar4Image *dst, ITMFloatImage 
 {
 	Vector4u *dest = dst->GetData(MEMORYDEVICE_CPU);
 	float *source = src->GetData(MEMORYDEVICE_CPU);
-	int dataSize = dst->dataSize;
+	int dataSize = static_cast<int>(dst->dataSize);
 	
 	float mindepth = 1000;
 	for (int i = 0; i < src->dataSize; i++)
